Added support for seamless iframes
Added an option for supporting seamless iframes. This assumes that the iframe is not scrollable and is meant to look as though it is part of the top window. In this type of iframe, all the elements are always visible. For that reason, we will detect whether elements are visible inside of the top window instead of the iframe.