Fix issue with seamless iframe support
Updated code for supporting seamless iframes to check if the iframe is
actually seamless before applying the logic. When the iframe is not
seamless, the logic should be the same as when the page is not in an
iframe at all.

These changes assume that seamless iframes will have scrolling="no" or
the HTML5 seamless attribute.